+++ src/main/java/com/avaje/ebeaninternal/server/autofetch/DefaultAutoFetchManager.java
@@ -527,7 +527,7 @@ public class DefaultAutoFetchManager implements AutoFetchManager, Serializable {
 	 * query in which case the parentNode will be null, or a lazy loading query
 	 * resulting from traversal of the object graph.
 	 */
-	public void collectQueryInfo(ObjectGraphNode node, int beans, int micros) {
+	public void collectQueryInfo(ObjectGraphNode node, long beans, long micros) {
 
 		if (node != null){
 			ObjectGraphOrigin origin = node.getOriginQueryPoint();
