+++ src/main/java/com/avaje/ebeaninternal/server/autofetch/StatisticsQuery.java
@@ -13,11 +13,11 @@ public class StatisticsQuery implements Serializable {
 
 	private final String path;
 	
-	private int exeCount;
+	private long exeCount;
 	
-	private int totalBeanLoaded;
+	private long totalBeanLoaded;
 	
-	private int totalMicros;
+	private long totalMicros;
 	
 	public StatisticsQuery(String path){
 		this.path = path;
@@ -27,7 +27,7 @@ public class StatisticsQuery implements Serializable {
 		return new QueryStats(path, exeCount, totalBeanLoaded, totalMicros);
 	}
 	
-	public void add(int beansLoaded, int micros) {
+	public void add(long beansLoaded, long micros) {
 		exeCount++;
 		totalBeanLoaded += beansLoaded;
 		totalMicros += micros;
